+++ CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.2.2] - 2026-08-17
+
+### Added
+
+- `force_percent` opt
+
+## [0.2.2] - 2026-08-18
+
+### Added
+
+- Lots of documentation and typespecs
+- more helper functions in `Curves.Utils.Predefined`
+
+### Changed
+
+Exposed some modules to documentation that were previously set to `@moduledoc false`.
+
+## [0.2.1] - 2026-08-17
+
+### Added
+
+- `Curves.take/3` and `Curves.take!/3`
+
 ## [0.2.0] - 2026-08-16
 
 A lot of preparatory work before implementing splines.